28.05

+++ Forces/Item.cs
@@ -2,7 +2,7 @@
 {
     public class Item
     {
-        public Item(Vector r, Vector speed, int mass, int volume)
+        public Item(Vector r, Vector speed, double mass, double volume)
         {
             R = r;
             Speed = Speed;
@@ -14,17 +14,17 @@
 
         public Vector Speed { get; set; }
 
-        public int Volume { get; set; }
+        public double Volume { get; set; }
 
-        public int Mass { get; set; }
+        public double Mass { get; set; }
 
-        public void Move(int dt, Vector F)
+        public void Move(double dt, Vector F)
         {
             Vector a = F / Mass;
             Speed += a * dt;
             Move(dt);
         }
 
-        public void Move(int dt) => R += Speed * dt;
+        public void Move(double dt) => R += Speed * dt;
     }
 }
